Split mill operators.test.ts into one test file per operator
operators.test.ts (76 tests) -> tests/operators/{set,unset,inc,mul,min,max,
rename,push,pop,pull,pull-all,add-to-set}.test.ts plus combinations.test.ts
and validation.test.ts for cross-operator/error cases. Each operator file holds
its behavior + reactivity tests. Pure relocation: every test moved verbatim,
189 tests still pass, 100% coverage.
