+++ packages/mill/tests/operators/set.test.ts
@@ -1,0 +1,49 @@
+import { createReactive, effect } from "@supergrain/kernel";
+import { describe, it, expect, vi } from "vitest";
+
+import { applyWithUndo } from "../helpers";
+
+describe("MongoDB Style Operators", () => {
+  it("$set: should set top-level and nested properties", () => {
+    const state = createReactive({
+      user: { name: "John", address: { city: "New York" } },
+    });
+    const { rewindAndAssertRestored } = applyWithUndo(
+      state,
+      {},
+      {
+        $set: { "user.name": "Jane", "user.address.city": "Boston" },
+      },
+    );
+    expect(state.user.name).toBe("Jane");
+    expect(state.user.address.city).toBe("Boston");
+    rewindAndAssertRestored();
+  });
+
+  it("$set creates missing nested paths", () => {
+    const store = createReactive<any>({});
+    const { undo, rewindAndAssertRestored } = applyWithUndo(
+      store,
+      {},
+      { $set: { "brand.new.path": "value" } },
+    );
+    expect(store.brand.new.path).toBe("value");
+    expect(undo).toEqual({ $unset: { brand: "" } });
+    rewindAndAssertRestored();
+  });
+
+  it("$set fires effects subscribed to the written path and not to siblings", () => {
+    const store = createReactive({ a: 1, b: 2 });
+    const aFn = vi.fn(() => void store.a);
+    const bFn = vi.fn(() => void store.b);
+    effect(aFn);
+    effect(bFn);
+
+    const { rewindAndAssertRestored } = applyWithUndo(store, {}, { $set: { a: 10 } });
+
+    expect(store.a).toBe(10);
+    expect(aFn).toHaveBeenCalledTimes(2);
+    expect(bFn).toHaveBeenCalledTimes(1);
+    rewindAndAssertRestored();
+  });
+});
